requirement.txt
faker>=18.0.0
pandas>=1.3.0
pillow>=9.0.0
numpy>=1.21.0
scikit-learn>=1.0.0
tensorflow>=2.6.0
librosa>=0.9.0
pyttsx3>=2.90
fastapi>=0.70.0
uvicorn>=0.16.0
python-multipart>=0.0.5